pyproject.toml
[project]
name = "working-time-tracker"
version = "0.1.0"
description = ""
authors = [
    { name = "osha890", email = "oshavshav@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "mypy (>=1.15.0,<2.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "django-stubs (>=5.2.0,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
mypy_path = "./working_time_tracker"

[tool.django-stubs]
django_settings_module = "working_time_tracker.settings"

[tool.isort]
profile = "black"
force_sort_within_sections = true
lines_between_sections = 1
lines_between_types = 1
known_first_party = ["tracker", "custom_auth"]
sections = ["STDLIB", "THIRDPARTY", "FIRSTPARTY"]

